If we have a query strign parameter like an API, because it was not passed to the swagger spec api call, the filters in the server side did not consider the api key and would filter ALL the routes of the API. The final effect was that none of the routes would appear.
So we propagated the query string (like an API key) when we call the swagger spec.

In this case, the url passed to the APIExplorerAPIConfigurator object will propagate the main site query string. This values will be passed to the call for the swagger spec and the filter in the server will consider the API key to filter the routes